Enforce fact-based linter rule helpers
