+++ crates/shuck-linter/src/rules/correctness/constant_comparison_test.rs
@@ -1,8 +1,8 @@
-use shuck_ast::{Span, Word, WordPart, WordPartNode};
+use shuck_ast::Span;
 
 use crate::{
     Checker, ConditionalNodeFact, ConditionalOperatorFamily, Rule, SimpleTestOperatorFamily,
-    SimpleTestShape, Violation,
+    SimpleTestShape, Violation, double_quoted_scalar_affix_span,
 };
 
 pub struct ConstantComparisonTest;
@@ -60,79 +60,7 @@ fn simple_test_is_constant(fact: &crate::SimpleTestFact<'_>) -> bool {
 fn simple_test_unary_affix_span(fact: &crate::SimpleTestFact<'_>) -> Option<Span> {
     let operand = fact.operands().get(1)?;
 
-    word_scalar_affix_span(operand)
-}
-
-fn word_scalar_affix_span(word: &Word) -> Option<Span> {
-    if !word.is_fully_double_quoted() {
-        return None;
-    }
-
-    let mut saw_literal = false;
-    let mut saw_scalar_expansion = false;
-    let mut literal_span = None;
-    if !word_scalar_affix_span_parts(
-        &word.parts,
-        &mut saw_literal,
-        &mut saw_scalar_expansion,
-        &mut literal_span,
-    ) {
-        return None;
-    }
-
-    (saw_literal && saw_scalar_expansion)
-        .then_some(literal_span)
-        .flatten()
-}
-
-fn word_scalar_affix_span_parts(
-    parts: &[WordPartNode],
-    saw_literal: &mut bool,
-    saw_scalar_expansion: &mut bool,
-    literal_span: &mut Option<Span>,
-) -> bool {
-    for part in parts {
-        match &part.kind {
-            WordPart::Literal(_) | WordPart::SingleQuoted { .. } => {
-                *saw_literal = true;
-                if literal_span.is_none() {
-                    *literal_span = Some(part.span);
-                }
-            }
-            WordPart::DoubleQuoted { parts, .. } => {
-                if !word_scalar_affix_span_parts(
-                    parts,
-                    saw_literal,
-                    saw_scalar_expansion,
-                    literal_span,
-                ) {
-                    return false;
-                }
-            }
-            WordPart::Variable(_)
-            | WordPart::Parameter(_)
-            | WordPart::ParameterExpansion { .. }
-            | WordPart::Length(_)
-            | WordPart::Substring { .. }
-            | WordPart::IndirectExpansion { .. }
-            | WordPart::Transformation { .. } => {
-                *saw_scalar_expansion = true;
-            }
-            WordPart::ArrayAccess(_)
-            | WordPart::ArrayLength(_)
-            | WordPart::ArrayIndices(_)
-            | WordPart::ArraySlice { .. }
-            | WordPart::PrefixMatch { .. }
-            | WordPart::CommandSubstitution { .. }
-            | WordPart::ArithmeticExpansion { .. }
-            | WordPart::ProcessSubstitution { .. }
-            | WordPart::ZshQualifiedGlob(_) => {
-                return false;
-            }
-        }
-    }
-
-    true
+    double_quoted_scalar_affix_span(operand)
 }
 
 fn simple_test_report_span(fact: &crate::SimpleTestFact<'_>, fallback: Span) -> Span {
