+++ crates/shuck-linter/src/rules/portability/conditional_portability.rs
@@ -1,11 +1,9 @@
-use shuck_ast::{
-    BourneParameterExpansion, Command, CompoundCommand, ConditionalBinaryOp, ConditionalCommand,
-    ConditionalExpr, ConditionalUnaryOp, ParameterExpansion, ParameterExpansionSyntax, Pattern,
-    PatternPart, Span, VarRef, Word, WordPart, WordPartNode, ZshExpansionTarget,
-};
+use shuck_ast::{ConditionalBinaryOp, ConditionalUnaryOp, Span};
 
 use crate::{
-    Checker, Rule, ShellDialect, SimpleTestFact, SimpleTestSyntax, Violation, static_word_text,
+    Checker, ConditionalNodeFact, Rule, ShellDialect, SimpleTestFact, SimpleTestSyntax, Violation,
+    conditional_array_subscript_span, conditional_extglob_span, static_word_text,
+    word_array_subscript_span, word_extglob_span,
 };
 
 pub struct DoubleBracketInSh;
@@ -172,7 +170,7 @@ pub fn double_bracket_in_sh(checker: &mut Checker) {
         .facts()
         .commands()
         .iter()
-        .filter(|fact| conditional_command(fact).is_some())
+        .filter(|fact| fact.conditional().is_some())
         .map(crate::CommandFact::span)
         .collect::<Vec<_>>();
 
@@ -213,12 +211,9 @@ pub fn test_equality_operator(checker: &mut Checker) {
             .facts()
             .commands()
             .iter()
-            .filter_map(conditional_command)
-            .flat_map(|command| {
-                conditional_binary_operator_spans(
-                    &command.expression,
-                    ConditionalBinaryOp::PatternEq,
-                )
+            .filter_map(|fact| fact.conditional())
+            .flat_map(|conditional| {
+                conditional_binary_operator_spans(conditional, ConditionalBinaryOp::PatternEq)
             }),
     );
 
@@ -235,7 +230,7 @@ pub fn if_elif_bash_test(checker: &mut Checker) {
         .commands()
         .iter()
         .filter(|fact| checker.facts().is_elif_condition_command(fact.id()))
-        .filter(|fact| conditional_command(fact).is_some())
+        .filter(|fact| fact.conditional().is_some())
         .map(crate::CommandFact::span)
         .collect::<Vec<_>>();
 
@@ -252,8 +247,8 @@ pub fn extended_glob_in_test(checker: &mut Checker) {
         .facts()
         .commands()
         .iter()
-        .filter_map(conditional_command)
-        .filter_map(|command| conditional_extglob_span(&command.expression, source))
+        .filter_map(|fact| fact.conditional())
+        .filter_map(|conditional| conditional_extglob_span(conditional.expression(), source))
         .collect::<Vec<_>>();
 
     checker.report_all_dedup(spans, || ExtendedGlobInTest);
@@ -291,8 +286,10 @@ pub fn array_subscript_condition(checker: &mut Checker) {
         .facts()
         .commands()
         .iter()
-        .filter_map(conditional_command)
-        .filter_map(|command| conditional_array_subscript_span(&command.expression, source))
+        .filter_map(|fact| fact.conditional())
+        .filter_map(|conditional| {
+            conditional_array_subscript_span(conditional.expression(), source)
+        })
         .collect::<Vec<_>>();
 
     checker.report_all_dedup(spans, || ArraySubscriptCondition);
@@ -329,12 +326,9 @@ pub fn greater_than_in_double_bracket(checker: &mut Checker) {
         .facts()
         .commands()
         .iter()
-        .filter_map(conditional_command)
-        .flat_map(|command| {
-            conditional_binary_operator_spans(
-                &command.expression,
-                ConditionalBinaryOp::LexicalAfter,
-            )
+        .filter_map(|fact| fact.conditional())
+        .flat_map(|conditional| {
+            conditional_binary_operator_spans(conditional, ConditionalBinaryOp::LexicalAfter)
         })
         .collect::<Vec<_>>();
 
@@ -350,9 +344,9 @@ pub fn regex_match_in_sh(checker: &mut Checker) {
         .facts()
         .commands()
         .iter()
-        .filter_map(conditional_command)
-        .flat_map(|command| {
-            conditional_binary_operator_spans(&command.expression, ConditionalBinaryOp::RegexMatch)
+        .filter_map(|fact| fact.conditional())
+        .flat_map(|conditional| {
+            conditional_binary_operator_spans(conditional, ConditionalBinaryOp::RegexMatch)
         })
         .collect::<Vec<_>>();
 
@@ -368,9 +362,9 @@ pub fn v_test_in_sh(checker: &mut Checker) {
         .facts()
         .commands()
         .iter()
-        .filter_map(conditional_command)
-        .flat_map(|command| {
-            conditional_unary_operator_spans(&command.expression, |operator, _| {
+        .filter_map(|fact| fact.conditional())
+        .flat_map(|conditional| {
+            conditional_unary_operator_spans(conditional, |operator, _| {
                 operator == ConditionalUnaryOp::VariableSet
             })
         })
@@ -389,9 +383,9 @@ pub fn a_test_in_sh(checker: &mut Checker) {
         .facts()
         .commands()
         .iter()
-        .filter_map(conditional_command)
-        .flat_map(|command| {
-            conditional_unary_operator_spans(&command.expression, |operator, op_span| {
+        .filter_map(|fact| fact.conditional())
+        .flat_map(|conditional| {
+            conditional_unary_operator_spans(conditional, |operator, op_span| {
                 operator == ConditionalUnaryOp::Exists && op_span.slice(source) == "-a"
             })
         })
@@ -409,9 +403,9 @@ pub fn option_test_in_sh(checker: &mut Checker) {
         .facts()
         .commands()
         .iter()
-        .filter_map(conditional_command)
-        .flat_map(|command| {
-            conditional_unary_operator_spans(&command.expression, |operator, _| {
+        .filter_map(|fact| fact.conditional())
+        .flat_map(|conditional| {
+            conditional_unary_operator_spans(conditional, |operator, _| {
                 operator == ConditionalUnaryOp::OptionSet
             })
         })
@@ -466,14 +460,6 @@ pub fn ownership_test_in_sh(checker: &mut Checker) {
 
 fn is_posix_sh_shell(shell: ShellDialect) -> bool {
     matches!(shell, ShellDialect::Sh | ShellDialect::Dash)
-}
-
-fn conditional_command<'a>(fact: &'a crate::CommandFact<'a>) -> Option<&'a ConditionalCommand> {
-    let Command::Compound(CompoundCommand::Conditional(command)) = fact.command() else {
-        return None;
-    };
-
-    Some(command)
 }
 
 fn simple_test_binary_operator_token_spans(
@@ -633,441 +619,42 @@ fn simple_test_command_span(
     Some(Span::from_positions(name.span.start, end))
 }
 
-fn conditional_extglob_span(expression: &ConditionalExpr, source: &str) -> Option<Span> {
-    match expression {
-        ConditionalExpr::Binary(expr) => conditional_extglob_span(&expr.left, source)
-            .or_else(|| conditional_extglob_span(&expr.right, source)),
-        ConditionalExpr::Unary(expr) => conditional_extglob_span(&expr.expr, source),
-        ConditionalExpr::Parenthesized(expr) => conditional_extglob_span(&expr.expr, source),
-        ConditionalExpr::Pattern(pattern) => pattern_extglob_span(pattern, source),
-        ConditionalExpr::Word(_) | ConditionalExpr::Regex(_) | ConditionalExpr::VarRef(_) => None,
-    }
-}
-
-fn pattern_extglob_span(pattern: &Pattern, source: &str) -> Option<Span> {
-    for part in &pattern.parts {
-        match &part.kind {
-            PatternPart::Group { patterns, .. } => {
-                return Some(part.span).or_else(|| {
-                    patterns
-                        .iter()
-                        .find_map(|pattern| pattern_extglob_span(pattern, source))
-                });
-            }
-            PatternPart::Word(word) => {
-                if let Some(span) = word_extglob_span(word, source) {
-                    return Some(span);
-                }
-            }
-            PatternPart::Literal(_)
-            | PatternPart::AnyString
-            | PatternPart::AnyChar
-            | PatternPart::CharClass(_) => {}
-        }
-    }
-
-    None
-}
-
-fn conditional_array_subscript_span(expression: &ConditionalExpr, source: &str) -> Option<Span> {
-    match expression {
-        ConditionalExpr::Binary(expr) => conditional_array_subscript_span(&expr.left, source)
-            .or_else(|| conditional_array_subscript_span(&expr.right, source)),
-        ConditionalExpr::Unary(expr) => conditional_array_subscript_span(&expr.expr, source),
-        ConditionalExpr::Parenthesized(expr) => {
-            conditional_array_subscript_span(&expr.expr, source)
-        }
-        ConditionalExpr::Word(word) | ConditionalExpr::Regex(word) => {
-            word_array_subscript_span(word, source)
-        }
-        ConditionalExpr::Pattern(pattern) => pattern_array_subscript_span(pattern, source),
-        ConditionalExpr::VarRef(reference) => var_ref_subscript_span(reference),
-    }
-}
-
-fn pattern_array_subscript_span(pattern: &Pattern, source: &str) -> Option<Span> {
-    for part in &pattern.parts {
-        match &part.kind {
-            PatternPart::Group { patterns, .. } => {
-                if let Some(span) = patterns
-                    .iter()
-                    .find_map(|pattern| pattern_array_subscript_span(pattern, source))
-                {
-                    return Some(span);
-                }
-            }
-            PatternPart::Word(word) => {
-                if let Some(span) = word_array_subscript_span(word, source) {
-                    return Some(span);
-                }
-            }
-            PatternPart::Literal(_)
-            | PatternPart::AnyString
-            | PatternPart::AnyChar
-            | PatternPart::CharClass(_) => {}
-        }
-    }
-
-    None
-}
-
-fn word_array_subscript_span(word: &Word, source: &str) -> Option<Span> {
-    word_array_subscript_span_from_parts(&word.parts, source).or_else(|| {
-        (!word.has_quoted_parts() && text_has_variable_subscript(word.span.slice(source)))
-            .then_some(word.span)
-    })
-}
-
-fn word_array_subscript_span_from_parts(parts: &[WordPartNode], source: &str) -> Option<Span> {
-    for part in parts {
-        match &part.kind {
-            WordPart::DoubleQuoted { parts, .. } => {
-                if let Some(span) = word_array_subscript_span_from_parts(parts, source) {
-                    return Some(span);
-                }
-            }
-            WordPart::Literal(_) => {
-                if text_has_variable_subscript(part.span.slice(source)) {
-                    return Some(part.span);
-                }
-            }
-            WordPart::Parameter(parameter) => {
-                if let Some(span) = parameter_array_subscript_span(parameter) {
-                    return Some(span);
-                }
-            }
-            WordPart::ParameterExpansion { reference, .. }
-            | WordPart::Length(reference)
-            | WordPart::ArrayAccess(reference)
-            | WordPart::ArrayLength(reference)
-            | WordPart::ArrayIndices(reference)
-            | WordPart::Substring { reference, .. }
-            | WordPart::ArraySlice { reference, .. }
-            | WordPart::IndirectExpansion { reference, .. }
-            | WordPart::Transformation { reference, .. } => {
-                if let Some(span) = var_ref_subscript_span(reference) {
-                    return Some(span);
-                }
-            }
-            WordPart::ZshQualifiedGlob(_)
-            | WordPart::SingleQuoted { .. }
-            | WordPart::Variable(_)
-            | WordPart::CommandSubstitution { .. }
-            | WordPart::ArithmeticExpansion { .. }
-            | WordPart::PrefixMatch { .. }
-            | WordPart::ProcessSubstitution { .. } => {}
-        }
-    }
-
-    None
-}
-
-fn parameter_array_subscript_span(parameter: &ParameterExpansion) -> Option<Span> {
-    match &parameter.syntax {
-        ParameterExpansionSyntax::Bourne(syntax) => match syntax {
-            BourneParameterExpansion::Access { reference }
-            | BourneParameterExpansion::Length { reference }
-            | BourneParameterExpansion::Indices { reference }
-            | BourneParameterExpansion::Indirect { reference, .. }
-            | BourneParameterExpansion::Slice { reference, .. }
-            | BourneParameterExpansion::Operation { reference, .. }
-            | BourneParameterExpansion::Transformation { reference, .. } => {
-                var_ref_subscript_span(reference)
-            }
-            BourneParameterExpansion::PrefixMatch { .. } => None,
-        },
-        ParameterExpansionSyntax::Zsh(syntax) => match &syntax.target {
-            ZshExpansionTarget::Reference(reference) => var_ref_subscript_span(reference),
-            ZshExpansionTarget::Nested(parameter) => parameter_array_subscript_span(parameter),
-            ZshExpansionTarget::Word(_) | ZshExpansionTarget::Empty => None,
-        },
-    }
-}
-
-fn var_ref_subscript_span(reference: &VarRef) -> Option<Span> {
-    reference
-        .subscript
-        .as_ref()
-        .filter(|subscript| subscript.selector().is_none())
-        .map(|_| reference.span)
-}
-
-fn word_extglob_span(word: &Word, source: &str) -> Option<Span> {
-    word_extglob_span_from_parts(&word.parts, source).or_else(|| {
-        (!word.has_quoted_parts()
-            && word_has_only_literal_parts(&word.parts)
-            && text_looks_like_extglob(word.span.slice(source)))
-        .then_some(word.span)
-    })
-}
-
-fn word_extglob_span_from_parts(parts: &[WordPartNode], source: &str) -> Option<Span> {
-    for part in parts {
-        match &part.kind {
-            WordPart::Literal(_) => {
-                if text_looks_like_extglob(part.span.slice(source)) {
-                    return Some(part.span);
-                }
-            }
-            WordPart::DoubleQuoted { .. } | WordPart::SingleQuoted { .. } => {}
-            WordPart::ZshQualifiedGlob(_)
-            | WordPart::Variable(_)
-            | WordPart::CommandSubstitution { .. }
-            | WordPart::ArithmeticExpansion { .. }
-            | WordPart::Parameter(_)
-            | WordPart::ParameterExpansion { .. }
-            | WordPart::Length(_)
-            | WordPart::ArrayAccess(_)
-            | WordPart::ArrayLength(_)
-            | WordPart::ArrayIndices(_)
-            | WordPart::Substring { .. }
-            | WordPart::ArraySlice { .. }
-            | WordPart::IndirectExpansion { .. }
-            | WordPart::PrefixMatch { .. }
-            | WordPart::ProcessSubstitution { .. }
-            | WordPart::Transformation { .. } => {}
-        }
-    }
-
-    None
-}
-
-fn word_has_only_literal_parts(parts: &[WordPartNode]) -> bool {
-    parts
-        .iter()
-        .all(|part| matches!(part.kind, WordPart::Literal(_)))
-}
-
-fn text_has_variable_subscript(text: &str) -> bool {
-    let bytes = text.as_bytes();
-    let mut index = 0usize;
-
-    while index < bytes.len() {
-        if bytes[index] != b'$' || byte_is_backslash_escaped(bytes, index) {
-            index += 1;
-            continue;
-        }
-
-        let next = index + 1;
-        if next >= bytes.len() {
-            break;
-        }
-
-        if bytes[next] == b'{' {
-            let mut cursor = next + 1;
-            while cursor < bytes.len() && bytes[cursor] != b'}' {
-                if bytes[cursor] == b'['
-                    && bytes[cursor + 1..].contains(&b']')
-                    && bytes[cursor + 1..].contains(&b'}')
-                {
-                    return true;
-                }
-                cursor += 1;
-            }
-            index = cursor.saturating_add(1);
-            continue;
-        }
-
-        if !is_name_start(bytes[next]) {
-            index += 1;
-            continue;
-        }
-
-        let mut cursor = next + 1;
-        while cursor < bytes.len() && is_name_continue(bytes[cursor]) {
-            cursor += 1;
-        }
-
-        if cursor < bytes.len() && bytes[cursor] == b'[' && bytes[cursor + 1..].contains(&b']') {
-            return true;
-        }
-
-        index = cursor;
-    }
-
-    false
-}
-
-fn text_looks_like_extglob(text: &str) -> bool {
-    let bytes = text.as_bytes();
-    if text_has_parenthesized_alternation(bytes) {
-        return true;
-    }
-
-    let mut index = 0usize;
-
-    while index + 1 < bytes.len() {
-        if !is_extglob_operator(bytes[index])
-            || bytes[index + 1] != b'('
-            || byte_is_backslash_escaped(bytes, index)
-        {
-            index += 1;
-            continue;
-        }
-
-        return matching_group_end(bytes, index + 1).is_some();
-    }
-
-    false
-}
-
-fn text_has_parenthesized_alternation(bytes: &[u8]) -> bool {
-    let mut index = 0usize;
-
-    while index < bytes.len() {
-        if bytes[index] != b'(' || byte_is_backslash_escaped(bytes, index) {
-            index += 1;
-            continue;
-        }
-
-        let Some(close) = matching_group_end(bytes, index) else {
-            index += 1;
-            continue;
-        };
-
-        if bytes[index + 1..close]
-            .iter()
-            .enumerate()
-            .any(|(offset, byte)| {
-                *byte == b'|' && !byte_is_backslash_escaped(bytes, index + 1 + offset)
-            })
-        {
-            return true;
-        }
-
-        index = close + 1;
-    }
-
-    false
-}
-
-fn matching_group_end(bytes: &[u8], open_index: usize) -> Option<usize> {
-    let mut depth = 1usize;
-    let mut cursor = open_index + 1;
-
-    while cursor < bytes.len() {
-        if byte_is_backslash_escaped(bytes, cursor) {
-            cursor += 1;
-            continue;
-        }
-
-        match bytes[cursor] {
-            b'(' => {
-                depth += 1;
-            }
-            b')' => {
-                depth -= 1;
-                if depth == 0 {
-                    return Some(cursor);
-                }
-            }
-            _ => {}
-        }
-
-        cursor += 1;
-    }
-
-    None
-}
-
-fn byte_is_backslash_escaped(bytes: &[u8], index: usize) -> bool {
-    let mut cursor = index;
-    let mut backslashes = 0usize;
-
-    while cursor > 0 && bytes[cursor - 1] == b'\\' {
-        backslashes += 1;
-        cursor -= 1;
-    }
-
-    backslashes % 2 == 1
-}
-
-fn is_extglob_operator(byte: u8) -> bool {
-    matches!(byte, b'@' | b'?' | b'+' | b'*' | b'!')
-}
-
 fn conditional_binary_operator_spans(
-    expression: &ConditionalExpr,
+    conditional: &crate::ConditionalFact<'_>,
     operator: ConditionalBinaryOp,
 ) -> Vec<Span> {
-    let mut spans = Vec::new();
-    collect_conditional_binary_operator_spans(expression, operator, &mut spans);
-    spans
-}
-
-fn collect_conditional_binary_operator_spans(
-    expression: &ConditionalExpr,
-    operator: ConditionalBinaryOp,
-    spans: &mut Vec<Span>,
-) {
-    match expression {
-        ConditionalExpr::Binary(expr) => {
-            if expr.op == operator {
-                spans.push(expr.op_span);
+    conditional
+        .nodes()
+        .iter()
+        .filter_map(|node| match node {
+            ConditionalNodeFact::Binary(binary) if binary.op() == operator => {
+                Some(binary.operator_span())
             }
-            collect_conditional_binary_operator_spans(&expr.left, operator, spans);
-            collect_conditional_binary_operator_spans(&expr.right, operator, spans);
-        }
-        ConditionalExpr::Unary(expr) => {
-            collect_conditional_binary_operator_spans(&expr.expr, operator, spans);
-        }
-        ConditionalExpr::Parenthesized(expr) => {
-            collect_conditional_binary_operator_spans(&expr.expr, operator, spans);
-        }
-        ConditionalExpr::Word(_)
-        | ConditionalExpr::Regex(_)
-        | ConditionalExpr::Pattern(_)
-        | ConditionalExpr::VarRef(_) => {}
-    }
+            _ => None,
+        })
+        .collect()
 }
 
 fn conditional_unary_operator_spans(
-    expression: &ConditionalExpr,
+    conditional: &crate::ConditionalFact<'_>,
     predicate: impl Fn(ConditionalUnaryOp, Span) -> bool + Copy,
 ) -> Vec<Span> {
-    let mut spans = Vec::new();
-    collect_conditional_unary_operator_spans(expression, predicate, &mut spans);
-    spans
-}
-
-fn collect_conditional_unary_operator_spans(
-    expression: &ConditionalExpr,
-    predicate: impl Fn(ConditionalUnaryOp, Span) -> bool + Copy,
-    spans: &mut Vec<Span>,
-) {
-    match expression {
-        ConditionalExpr::Binary(expr) => {
-            collect_conditional_unary_operator_spans(&expr.left, predicate, spans);
-            collect_conditional_unary_operator_spans(&expr.right, predicate, spans);
-        }
-        ConditionalExpr::Unary(expr) => {
-            if predicate(expr.op, expr.op_span) {
-                spans.push(expr.op_span);
+    conditional
+        .nodes()
+        .iter()
+        .filter_map(|node| match node {
+            ConditionalNodeFact::Unary(unary) if predicate(unary.op(), unary.operator_span()) => {
+                Some(unary.operator_span())
             }
-            collect_conditional_unary_operator_spans(&expr.expr, predicate, spans);
-        }
-        ConditionalExpr::Parenthesized(expr) => {
-            collect_conditional_unary_operator_spans(&expr.expr, predicate, spans);
-        }
-        ConditionalExpr::Word(_)
-        | ConditionalExpr::Regex(_)
-        | ConditionalExpr::Pattern(_)
-        | ConditionalExpr::VarRef(_) => {}
-    }
+            _ => None,
+        })
+        .collect()
 }
 
 #[derive(Clone, Copy, PartialEq, Eq)]
 enum SimpleTestOperatorKind {
     Unary,
     Binary,
-}
-
-fn is_name_start(byte: u8) -> bool {
-    byte == b'_' || byte.is_ascii_alphabetic()
-}
-
-fn is_name_continue(byte: u8) -> bool {
-    is_name_start(byte) || byte.is_ascii_digit()
 }
 
 #[cfg(test)]
